+++ packages/editor/src/RichContentEditor/Toolbars/createEditorToolbars.js
@@ -1,5 +1,4 @@
-import merge from 'lodash/merge';
-import { simplePubsub, getToolbarTheme, getConfigByFormFactor, TOOLBARS } from 'wix-rich-content-common';
+import { simplePubsub, getToolbarTheme, getConfigByFormFactor, TOOLBARS, DISPLAY_MODE, mergeToolbarSettings } from 'wix-rich-content-common';
 import { getDefaultToolbarSettings } from './default-toolbar-settings';
 import { MobileTextButtonList, DesktopTextButtonList } from './buttons';
 import { reducePluginTextButtonNames, mergeButtonLists } from './buttons/utils';
@@ -40,30 +39,21 @@ const createEditorToolbars = config => {
     desktop: mergeButtonLists(DesktopTextButtonList, reducePluginTextButtonNames(pluginTextButtons), 'desktop', appendSeparator)
   };
 
-  const defaultToolbarSettings = getDefaultToolbarSettings({ pluginButtons, textButtons, pluginTextButtons });
+  const defaultSettings = getDefaultToolbarSettings({ pluginButtons, textButtons, pluginTextButtons });
   const customSettings = getToolbarSettings({ pluginButtons, textButtons, pluginTextButtons });
-
-  const toolbarSettings = defaultToolbarSettings.reduce((mergedSettings, defaultSetting) => {
-    const customSettingsByName = customSettings.filter(s => s.name === defaultSetting.name);
-    if (customSettingsByName.length > 0) {
-      mergedSettings.push(merge(defaultSetting, customSettingsByName[0]));
-    } else {
-      mergedSettings.push(defaultSetting);
-    }
-    return mergedSettings;
-  }, []);
-
+  const toolbarSettings = mergeToolbarSettings({ defaultSettings, customSettings });
   const toolbars = {};
 
   toolbarSettings
     .filter(({ name }) => name !== TOOLBARS.PLUGIN)
     .filter(({ shouldCreate }) => getConfigByFormFactor({ config: shouldCreate(), isMobile, defaultValue: true }))
-    .forEach(({ name, getButtons, getTextPluginButtons, getPositionOffset, getVisibilityFn, getInstance }) => {
+    .forEach(({ name, getButtons, getTextPluginButtons, getPositionOffset, getVisibilityFn, getInstance, getDisplayOptions }) => {
       toolbars[name] = getInstance({
         buttons: getConfigByFormFactor({ config: getButtons(), isMobile, defaultValue: [] }),
         textPluginButtons: getConfigByFormFactor({ config: getTextPluginButtons(), isMobile, defaultValue: [] }),
         offset: getConfigByFormFactor({ config: getPositionOffset(), isMobile, defaultValue: { x: 0, y: 0 } }),
         visibilityFn: getConfigByFormFactor({ config: getVisibilityFn(), isMobile, defaultValue: () => true }),
+        displayOptions: getConfigByFormFactor({ config: getDisplayOptions(), isMobile, defaultValue: { displayMode: DISPLAY_MODE.NORMAL } }),
         theme: { ...getToolbarTheme(theme, name.toLowerCase()), ...theme },
         defaultTextAlignment: textAlignment,
         getEditorState,